+++ extensions/browser-bookmarks/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Update extension icon to match Raycast icons style and avoid any mismatch with "Day One" app
 
+## [Support for Chrome Beta Browser] - 2023-09-30
+
+- Add support for `Chrome Beta` browser
+
 ## [Support for Vivaldi and Edge Dev Browsers] - 2023-07-30
 
 - Add support for `Vivaldi`, `Edge Dev` and `Edge Canary` browsers